Override the channel platform arch
The following ensures that what you request is what you get back if we
end up in a situation where we forced the architecture from the API.
